Get rid of some compiler warnings

+++ include/util/utils.h
@@ -16,6 +16,6 @@ uint16_t convertEndian16(uint16_t val);
 // convert an integer from LSB to MSB and vice versa
 uint32_t convertEndian32(uint32_t val);
 
-#define SIZEOF_ARRAY( a ) (sizeof( a ) / sizeof( a[ 0 ] ))
+#define SIZEOF_ARRAY( a ) (int)(sizeof( a ) / sizeof( a[ 0 ] ))
 
 #endif /* UTILS_H_ */
